Safari-609.1.11.1/LayoutTests/platform/ios/media/video-controls-captions-trackmenu-localized-expected.txt

Test that captions and subtitles menu items are localized.


EVENT(canplaythrough)

*** Set the user language preference.
RUN(internals.setUserPreferredLanguages(['en']))
clickCCButton failed with exception: "undefined: FAIL
END OF TEST
